requirements.txt
pandas
python-dotenv
SQLAlchemy
pyodbc
chardet
